requirements.txt
absl-py==2.2.2
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
narwhals==1.36.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
setuptools==79.0.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
